Moooore refactors
