+++ src/client/features/login/components/login.tsx
@@ -1,12 +1,21 @@
 import React from 'react';
-import {
-  LoginContainer,
-  StyledGoogleButton,
-} from '../../team/components/styled';
+import GoogleButton from 'react-google-button';
+import styled from 'styled-components';
 
 interface Props {
   signIn: () => void;
 }
+
+const LoginContainer = styled.div`
+  padding: 15px;
+  display: flex;
+  width: 100%;
+  height: 50vh;
+`;
+
+const StyledGoogleButton = styled(GoogleButton)`
+  margin: auto;
+`;
 
 export const Login = ({ signIn }: Props) => (
   <LoginContainer>
