+++ src/client/features/shared/index.ts
@@ -1,8 +1,10 @@
-export { Icon } from './components/icon';
-export { Loader } from './components/loader';
 export { AlertContainer } from './alert-banner/components/alert-container';
 export { useSendAlert } from './alert-banner/hooks/use-send-alert';
 export {
   alertReducer,
   initialState as initialAlertState,
 } from './alert-banner/state/alert-slice';
+export { GolfersListItem } from './components/golfers-list-item';
+export { Icon } from './components/icon';
+export { Loader } from './components/loader';
+export { useManageGolfers } from './hooks/use-manage-golfers';
