+++ src/client/features/team-page/components/selected-team.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { useGetGolferData } from 'src/client/data/hooks/use-get-golfer-data';
 import { selectPhaseSelection } from 'src/client/store';
-import { useManageGolfers } from '../shared/hooks/use-manage-golfers';
-import { GolfersListItem } from '../team/components/golfers-list-item';
-import { StyledGolfersList, TeamContainer } from '../team/components/styled';
+import { StyledGolfersList } from '../../available-golfers/components/styled';
+import { GolfersListItem, useManageGolfers } from '../../shared';
 import { Name } from './name';
 import { SelectionControls } from './selection-controls';
+import { TeamContainer } from './styled';
 
 export const SelectedTeam = () => {
   const { getGolfersData } = useGetGolferData();
