+++ src/client/features/team-page/index.ts
@@ -1,4 +1,3 @@
-export { useManageGolfers } from '../shared/hooks/use-manage-golfers';
 export { TeamPage } from './components/team-page';
 export {
   currentTeamReducer,
